colorspace: generalize LMS model and make user-configurable
Mostly done for the purposes of re-using some of these functions
internally with different LMS models, but also in the hopes that some
users who specifically want the Bradford matrix for compatibility
reasons can influence this decision.

We continue using the CAT97 model by default since it's designed for use
with a linear transformation, which is exactly what we're doing in our
gamut mapping math.