(WIP) dispatch: implement dispatch commands
Recreating the shaders on every single frame is obviously a bit
wasteful. While this does amount to simply looking it up in the cache,
and seems for the most part unproblematic on modern systems, it does
represent pointless CPU load - and some clients may not want to keep
around all the state needed to recreate the specific shader they want to
re-execute multiple times. It may also be relevant for battery lifetime
on mobile.

It's up to the user to ensure the validity of any objects referenced by
the dispatch graph, as well as compatibility between dynamic descriptor
objects and the objects the shader was generated against. In the
pl_renderer I will most likely do a hard memcmp of the various params
structs and destroy the dispatch graph if anything at all has changed.

Closes https://code.videolan.org/videolan/libplacebo/-/issues/98