(WIP) shaders: automatically ensure identifier uniqueness
https://code.videolan.org/videolan/libplacebo/-/issues/106

The remaining issue with this approach is that the `pl_dispatch` code
still accesses raw identifiers *before* they're alpha-renamed (as part of
`pl_shader_finalize`). Either we need to violate even more abstractions
by having `pl_dispatch` also do renaming, or we drop this approach
altogether.

I'm not even sure what problem this is solving. What a waste of time.