vulkan: support pseudo-HDR10 mode via VK_EXT_hdr_metadata
This is not an optimal solution by any stretch of the imagination.
Instead it's a hack, meant to work around what I believe to be a driver
bug in the current implementations (affecting both AMD and Nvidia,
apparently).

Basically, the idea is that for implementations which support
VK_EXT_hdr_metadata but not VK_EXT_swapchain_colorspace, we can use the
former to force the display into HDR10 mode. This means we should
advertise support for such formats, on these platforms.

Closes https://code.videolan.org/videolan/libplacebo/-/issues/152