(WIP) rotation
This is not actually correct. We actually need to apply the rotation as
part of the samplers, rather than doing it via the output pass. This is
because the samplers need to know what the sampling ratio in each
direction is, and rotated sampling implies swapping the width/height
when calculating the scaling ratio.

Annoying, but needs to be redone.

cf. #144