shaders/colorspace: support pl_color_repr in pl_shader_color_map
The primary reason this is done is so we can tone-map premultiplied
alpha content properly. But as an aside, this means that it now also
supports tone mapping e.g. YUV content, which we may be able to do more
elegantly in the future. (Currently, this is just done naively by
transforming to RGB - but in principle, we could adopt some of the
smarter YUV-only tone mapping algorithms)

Also rewrite the function a bit to make it less confusing to think
about. (And less error-prone)

Note that, due to an unfortunate implementation quirk, this will
actually first multiply and *then* again un-multiply content that is
currently independent alpha. Then again, this may be a feature, rather
than a bug. (Since it explicitly clears any alpha=0.0 content)

Anyway, this makes it so that color mapping is always done in
alpha-independent space, which fixes bugs related to overlay blending.