shaders/colorspace: use new tone mapping API
For convenience, I refactored the peak detection state object to hold
the LUT state as well. This avoids having to break the API, and
semantically groups the two together in a way that makes sense and
provides a straightforward amount of backwards compatibility.

For peak detection to work, we generate a 2D matrix of static tone
mapping curves, one for each possible peak value, and sample from the
one matching the detected peak at runtime.

Also apply the necessary (minor) changes to pl_renderer.

Other notable changes / omissions:
- `max_boost` can no longer really be used, which also means that we
  currently completely ignore `sig_avg` in the tone mapping logic. I may
  reintroduce this at some point if demand is high enough, but I need to
  figure out a clean/elegant way of integrating it.

- The old desaturation logic was replaced due to the existence of the
  new "luma" gamut mapping mode, which basically does the same thing as
  max+desat except better. Again, this may be reintroduced if demand is
  high enough, but for now I want to keep the code as simple as
  possible.

Closes https://code.videolan.org/videolan/libplacebo/-/issues/172