shaders/colorspace: simplify gamut clipping logic
This was way too complicated. Not only do we not care about the luma
coefficients, but we also don't even need to go through hoops to
calculate the right amount to add. When you think about it, it literally
just is straightforward: we always want to add the same amount to all
channels, after all (since 1/1/1 is neutral luma by definition in RGB),
and we know that we need to get the smallest channel in-range, so we
just literally add that amount.