shaders/colorspace: disable gamut clipping by default
This has some unwanted side effects in a lot of cases, as closer
investigation revealed. In particular, it has a tendency to destroy very
dark out-of-gamut details (due to hard-clipping all colors to in-gamut
equivalents). It also isn't very perceptually linear, because of issues
inherent to RGB color spaces. Easy enough to keep it around, but simply
disable it for now.